rename components
renamed a few components to simplify code.
